+++ app/blog/[slug]/page.tsx
@@ -1,48 +1,39 @@
-"use client"
-import { PostsConsumer } from "contexts/PostsContext"
-import { BackButton } from "components/ui/BackButton"
-import useWindowTop from "hooks/useWindowTop"
-import { authorName, textLogo } from "utils/constants"
-import MarkdownCode from "components/MarkdownCode"
-
-interface Params {
-    slug: string
+import { Metadata, ResolvingMetadata } from 'next'
+import Post from 'app/blog/[slug]/post'
+import { Post as PostInterface } from 'utils/interfaces'
+import supabase from 'utils/supabase'
+import { authorName, languages, textLogo } from 'utils/constants'
+ 
+type Props = {
+  params: { slug: string }
+  searchParams: { [key: string]: string | string[] | undefined }
 }
-
-interface props {
-    params: Params
+ 
+export async function generateMetadata ({ params, }: Props, ): Promise<Metadata> {
+  const slug = params.slug
+  const languagesCode = languages.map( lng => lng.code)
+  const { data, error } = await supabase
+  .from('Posts')
+  .select()
+  .eq("lang", languagesCode[1])
+  .eq("slug", slug)
+  const post = (data as PostInterface[])[0]
+  return {
+    description: `Article about: ${post?.title ?? "programming"}, written by ${authorName} (${textLogo.name}${textLogo.lastName}), web and iOS developer.`,
+    openGraph: {
+      title: post?.title ?? "",
+      description: `written by ${authorName} (${textLogo.name}${textLogo.lastName})`,
+      images: [
+        post?.image ?? ""
+      ]
+    },
+    twitter: {
+      title: post?.title ?? "",
+      description: `written by ${authorName} (${textLogo.name}${textLogo.lastName})`,
+      images: [
+        post?.image ?? ""
+      ]
+    }
+  }
 }
-
-const Post: React.FunctionComponent<props> = ({params})  => {
-  useWindowTop()
-  const post = PostsConsumer().filter( ({slug}) => slug === params.slug)[0]
-  return (
-    <>
-    <title>{`${post?.title ?? "Blog"} - ${textLogo.name}${textLogo.lastName}`}</title>
-    <meta name="description" content={`Article about: ${post?.title ?? "programming"}, written by ${authorName} (${textLogo.name}${textLogo.lastName}), web and iOS developer.`}/>
-    <meta property="og:url" content=""/>
-    <meta property="og:image" content={post?.image}/>
-    <meta property="og:title" content={post?.title}/>
-    <meta property="og:description" content={`written by ${authorName} (${textLogo.name}${textLogo.lastName})`}/>
-    <meta name="twitter:card" content="summary_large_image"/>
-    <meta property="twitter:url" content=""/>
-    <meta name="twitter:title" content={post?.title}/>
-    <meta property="twitter:description" content={`written by ${authorName} (${textLogo.name}${textLogo.lastName})`}/>
-    <meta name="twitter:image" content={post?.image}/>
-    <section className="max-w-7xl mx-auto overflow-hidden pt-32 pb-10 px-5 md:px-10 w-full">
-      <BackButton/>
-      {
-        post &&
-        <>
-          <h2 className="font-bold text-5xl text-center mt-10 py-5">{post?.title}</h2>
-          <div className="prose max-w-3xl dark:prose-invert mx-auto">
-          <MarkdownCode markdown={post.markdown}/>
-          </div>
-        </>
-      }
-    </section>
-    </>
-  )
-}
-
 export default Post
